genai-image-generation/requirements.txt
dify_plugin~=0.0.1b67
google-genai==1.2.0